requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
supabase==2.4.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
httpx==0.27.0
qrcode==7.4.2
Pillow==10.3.0
anthropic==0.25.0
pydantic==2.7.0
pydantic-settings==2.2.1
python-multipart==0.0.9
email-validator==2.1.1
